Refactoring on animationQueue + inline animations
The list is now completely flat, and you can use the start parameter on
Animation objects to determine whether it will start "onstep" or
"withprevious".
Inline animations can be defined by wrapping code in { } inside the
attribute. It is safe to assume that elem, context, params and callback
will all be set correctly