Build in gitattributes, animation refinements
Now the build directory acts as binary to avoid serious merge conflicts
(just rebuild after a merge), slide animations now work properly (moved
more parameters to the OliverAndSwan object). You can also pass the
start parameter in the parameters of an animation and it will
automatically put it to the Animation object.