Presenter view, mini test server, event handling
Everything to make this a more robust library. The test server was to
get around Chrome's strict policies with local html files, to open it
just run node swanTest.js
Event handling is so the presenter view can listen for events in the
presentation
Presenter view is just an html document with a ready function that takes
an argument of an OliverAndSwan presentation. It can retrieve things
like jQuery from the parent window (having two distinct jQuerys could be
messy)